tests/wpt/tests/css/css-gaps/animation/gap-decorations-width-neutral-keyframe-004.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>gap decorations column-rule-width neutral keyframe</title>
    <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#column-column-rule-width">
    <meta name="assert" content="gap decorations column-rule-width with repeaters supports neutral keyframe.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/support/interpolation-testcommon.js"></script>
  </head>
  <body>
    <div id="target"></div>
    <script>
      test(() => {
        target.style.columnRuleStyle = 'solid';
        target.style.columnRuleWidth= 'repeat(2, 20px 20px) 30px 30px';
        var animation = target.animate([{}, {columnRuleWidth: 'repeat(2, 40px 40px) 40px 40px'}], 1000);
        animation.pause();
        animation.currentTime = 500;
        assert_equals(getComputedStyle(target).columnRuleWidth, 'repeat(2, 30px 30px) 35px 35px');
      }, 'gap decorations column-rule-width with repeaters supports neutral keyframe.');
    </script>
  </body>
</html>
